% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{eq_clean_data}
\usage{
eq_clean_data(adf)
}
\arguments{
\item{adf}{a dataframe containing NOAA earthquake data}
}
\value{
This function returns a dataframe with latitude and longitude columns as numeric type
   and a  additional column date
}
\description{
This is a function the converts the latitude and longitude from NOAA earthquake data from 
   character to numeric and creates date column from YEAR,MONTH,DAY.
}
\examples{
\dontrun{
eq_clean_data(mydata)
}

}
