% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_create_label}
\alias{eq_create_label}
\title{eq_label}
\usage{
eq_create_label(adf)
}
\arguments{
\item{adf}{A dataframe containing NOAA earthquake data}
}
\value{
This function returns character text  formatted in html.
}
\description{
This function generates pop text in html format, give NOAA earthquake data.
}
\examples{
\dontrun{
eq_label(mydata)
}

}
