% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{eq_location_clean}
\usage{
eq_location_clean(adf)
}
\arguments{
\item{adf}{a dataframe containing NOAA earthquake data}
}
\value{
This function returns a dataframe with location_name column as formated as title case without the country name.
}
\description{
This is a function the converts the location_name from NOAA earthquake data from uppercase to title case and remove the country from location_name. (removes everything up to the last colon character.)
}
\examples{
eq_location_clean(mydata)

}
