% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_map}
\alias{eq_map}
\title{eq_map}
\usage{
eq_map(adf, annot_col)
}
\arguments{
\item{adf}{A dataframe containing NOAA earthquake data}

\item{annot_col}{A text string which is the name of the column in adf that 
contains the "pop up" text for the data points.
This "pop text" can be generated using the function eq_create_label.}
}
\value{
This function returns a !!need to fill
}
\description{
This function generates a leaflet map, give NOAA earthquake data.
}
\examples{
\dontrun{
eq_map(mydata,"popup_text")
}

}
