% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\name{geom_timeline}
\alias{geom_timeline}
\title{geom_timeline}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", show.legend = NA, na.rm = FALSE,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by ‘aes’. Can be inherited from 
upper levels of the plot.}

\item{data:}{The data to be displayed in this layer. Can be NULL, data.frame or function}
}
\description{
This function generates a timeline plot with years as y axis and individual earthquake 
   dates as points on the x axis.
}
\details{
: aes parameters act similar to `geom_point`
}
\section{Aesthetics}{

\aesthetics{geom}{timeline}
}

\examples{
\dontrun{
ggplot(quakesFiltered,aes(x = YEAR, y = 1,
        xmin = minYear,xmax = maxYear,
        )) +
 geom_timeline(aes(size = EQ_PRIMARY))
}


}
