% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{geom_timeline_label}
\usage{
geom_timeline_label(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", show.legend = NA, na.rm = FALSE,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by ‘aes’. Can be inherited from 
upper levels of the plot.}

\item{data:}{The data to be displayed in this layer. Can be NULL, data.frame or function}
}
\description{
This function generates annotation to accompany `geom_timeline` plot objects.
   The name of the earthquake location is shown above the timeline point and a
   line connection the name with the point is drawn.
}
\details{
: aes parameters act similar to `geom_label`
}
\examples{
\dontrun{
ggplot(quakesFiltered,aes(x = YEAR, y = 1,
        xmin = minYear,xmax = maxYear)) +
 geom_timeline(aes(size = EQ_PRIMARY)) +
 geom_timeline_label(aes(size = EQ_PRIMARY)) +
}


}
