% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\name{geom_timeline}
\alias{geom_timeline}
\title{geom_timeline}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", show.legend = NA, na.rm = FALSE,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by ‘aes’. Can be inherited from 
upper levels of the plot.}

\item{data}{The data to be displayed in this layer. Can be NULL, data.frame or function}

\item{stat}{The statistical transformation to use on the data for this layer}

\item{position}{Position adjustment, either as a string, or the result of a
call to a position adjustment function.}

\item{show.legend}{logical. Should this layer be included in the legends?}

\item{na.rm}{If ‘FALSE’, the default, missing values are removed with a
warning. If ‘TRUE’, missing values are silently removed.}

\item{inherit.aes}{If ‘FALSE’, overrides the default aesthetics, rather than
combining with them.}

\item{...}{other arguments passed on to ‘layer’.}
}
\description{
This function generates a timeline plot with years as y axis and individual earthquake 
   dates as points on the x axis.
}
\details{
: aes parameters act similar to `geom_point`
}
\examples{
\dontrun{
ggplot(quakesFiltered,aes(x = YEAR, y = 1,
        xmin = minYear,xmax = maxYear,
        )) +
 geom_timeline(aes(size = EQ_PRIMARY))
}


}
