% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{geom_timeline_label}
\usage{
geom_timeline_label(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", show.legend = NA, na.rm = FALSE,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by ‘aes’. Can be inherited from 
upper levels of the plot.}

\item{data}{The data to be displayed in this layer. Can be NULL, data.frame or function}

\item{stat}{The statistical transformation to use on the data for this layer}

\item{position}{Position adjustment, either as a string, or the result of a
call to a position adjustment function.}

\item{show.legend}{logical. Should this layer be included in the legends?}

\item{na.rm}{If ‘FALSE’, the default, missing values are removed with a
warning. If ‘TRUE’, missing values are silently removed.}

\item{inherit.aes}{If ‘FALSE’, overrides the default aesthetics, rather than
combining with them.}

\item{...}{other arguments passed on to ‘layer’.}
}
\description{
This function generates annotation to accompany `geom_timeline` plot objects.
   The name of the earthquake location is shown above the timeline point and a
   line connection the name with the point is drawn.
}
\details{
aes parameters act similar to `geom_label`
    aes 
\itemize{
\item{ x: a Date object}
\item{ y: single integer or a factor. (points will be grouped by y) example: COUNTRY}
\item{n_max: integer max number of labels sorted by size }
\item{size: numeric data field to base limiting of number of labels }
\item{label: character text to be diplayed over points }

}
}
\examples{
\dontrun{
ggplot(quakesFiltered,aes(x = DATE, y = COUNTRY,
        xmin = as.Date("2000-1-1","\%Y-\%m-\%d"),
        xmax = as.Date("2015-12-31","\%Y-\%m-\%d"),
        )) +
 geom_timeline(aes(size = EQ_PRIMARY)) +
 geom_timeline_label(aes(size = EQ_PRIMARY)) +
}


}
